requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
awsebcli==3.25
bcrypt==4.3.0
blessed==1.21.0
boto3==1.39.14
botocore==1.39.14
cement==2.10.14
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.4
django-appconf==1.1.0
django-cryptography==1.1
django-fernet-fields==0.6
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-storages==1.14.6
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
fabric==3.2.2
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
multidict==6.6.3
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
phonenumbers==9.0.10
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.4
s3transfer==0.13.1
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
twilio==9.7.0
typing_extensions==4.14.1
urllib3==1.26.20
wcwidth==0.2.13
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
